Moved test to PostStoreUnitTest
